add machinist gauge updates also fix some potencies/guaranteed hits start test skeletons for other jobs
