+++ fflogs_rotation/job_data/guaranteed_hits_by_action.csv
@@ -6,3 +6,4 @@ action_id,action_name,hit_type,valid_start,valid_end
 25782,Kaeshi: Namikiri,1,1661144400000,2733119200000
 25781,Ogi Namikiri,1,1661144400000,2733119200000
 25792,Starfall Dance,3,1661144400000,2733119200000
+36982,Full Metal Field,3,1719565200000,2733119200000
