+++ fflogs_rotation/job_data/potencies/physical_ranged/Machinist/7_01-Machinist.csv
