+++ fflogs_rotation/tests/test_machinist.py
@@ -1,0 +1,2 @@
+from fflogs_rotation.machinist import MachinistActions
+
