+++ fflogs_rotation/tests/test_monk.py
@@ -1,0 +1,40 @@
+from fflogs_rotation.monk import MonkActions
+
+# TODO: set up monk test
+
+# Stats
+# HP
+# 122060
+# Strength
+# 4284
+# Vitality
+# 4349
+# Critical Hit
+# 2208
+# Direct Hit Rate
+# 1976
+# Determination
+# 1897
+# Skill Speed
+# 964
+# https://www.fflogs.com/reports/gVkPC9HXaz7tjWbn#fight=1&type=damage-done
+# https://www.fflogs.com/reports/gVkPC9HXaz7tjWbn#fight=4
+# RotationTable(
+#     headers,
+#     "gVkPC9HXaz7tjWbn",
+#     5,
+#     "Monk",
+#     59,
+#     2208,
+#     1976,
+#     1897,
+#     351,
+#     100,
+#     damage_buff_table,
+#     critical_hit_rate_table,
+#     direct_hit_rate_table,
+#     guaranteed_hits_by_action_table,
+#     guaranteed_hits_by_buff_table,
+#     potency_table,
+#     pet_ids=None,
+# )
