+++ requirements.txt
@@ -3,7 +3,7 @@ dash-bootstrap-components==1.6.0
 dash-core-components==2.0.0
 dash-html-components==2.0.0
 dash-table==5.0.0
-ffxiv_stats==0.5.0
+ffxiv_stats==0.5.2
 Flask==3.0.0
 numpy==1.26.4
 pandas==2.0.0
